+++ src/textstring/TextStringComplex.tsx
@@ -3,7 +3,7 @@ import React, { FunctionComponent, memo, ReactElement, ReactNode, useEffect, use
 import { TextString } from 'chayns-components';
 // @ts-expect-error
 import isTobitEmployee from 'chayns-components/dist/esm/utils/tobitEmployee.js';
-import generateUUID from '../functions/generateGuid';
+import generateGuid from '../functions/generateGuid';
 import jsxReplace, { JsxReplacements } from './jsxReplace';
 import TEXTSTRING_CONFIG from './textstringConfig';
 import isNullOrWhiteSpace from '../utils/isNullOrWhiteSpace';
@@ -17,24 +17,6 @@ export interface TextStringComplexConfig {
     useDangerouslySetInnerHTML?: boolean;
     language?: string;
     autoCreation?: boolean;
-}
-
-const searchTextString = (
-    searchKey: string,
-    stringContainer: string
-        | Record<string, string>
-        | Record<string, Record<string, string>>
-        | Record<string, Record<string, Record<string, string>>>
-): boolean => {
-    const keys = Object.keys(stringContainer);
-    if (keys.includes(searchKey)) return true;
-    return (keys as unknown[] as boolean[])
-        .reduce(
-            (total, current) =>
-                total ||
-                searchTextString(searchKey, (stringContainer as Record<string, any>)[current as unknown as string]),
-            false
-        );
 }
 
 /**
@@ -59,75 +41,71 @@ const TextStringComplex: FunctionComponent<TextStringComplexConfig> = ({
     maxReplacements = 20,
     useDangerouslySetInnerHTML = false,
     language = undefined,
-    autoCreation = process.env.NODE_ENV === 'production',
+    autoCreation: propAutoCreation,
     ...props
 }) => {
-    // create missing textStrings in QA/Production if opened by an authorized developer
+    const autoCreation = propAutoCreation ?? TEXTSTRING_CONFIG.autoCreation;
+    // create missing textStrings if opened by an authorized developer
     useEffect(() => {
         (async () => {
             try {
                 if (autoCreation
                     && !isNullOrWhiteSpace(fallback)
                     && !isNullOrWhiteSpace(stringName)
-                    && (TextString.getTextString(stringName) ?? null) === null
+                    && (TextString.getTextString(`${TEXTSTRING_CONFIG.prefix}${stringName}`) ?? null) === null
+                    && (TextString.getTextString(`${TEXTSTRING_CONFIG.prefix}${stringName}`, 'de') ?? null) === null
                     && !isNullOrWhiteSpace(TEXTSTRING_CONFIG.libName)
                     && !isNullOrWhiteSpace(TEXTSTRING_CONFIG.prefix)
                     && chayns.env.user.isAuthenticated
                 ) {
-                    if (!searchTextString(`${TEXTSTRING_CONFIG.prefix}${stringName}`, TextString.textStrings)) {
-                        isTobitEmployee().then(async () => {
-                            const libResponse = await fetch(
-                                `https://webapi.tobit.com/TextStringService/v1.0/V2/LangLibs/${TEXTSTRING_CONFIG.libName}`,
+                    isTobitEmployee().then(async () => {
+                        const libResponse = await fetch(
+                            `https://webapi.tobit.com/TextStringService/v1.0/langstrings/${TEXTSTRING_CONFIG.libName}?language=de`,
+                            {
+                                method: 'GET',
+                                cache: 'no-cache',
+                                headers: new Headers({
+                                    Authorization: `Bearer ${chayns.env.user.tobitAccessToken}`
+                                })
+                            }
+                        );
+                        const libContent = await libResponse.json();
+                        if (libResponse.status === 200 && libContent &&
+                            Object.prototype.toString.call(libContent) === '[object Object]' &&
+                            Object.keys(libContent)
+                                .find(name => name === `${TEXTSTRING_CONFIG.prefix}${stringName}`)
+                        ) {
+                            const response = await fetch(
+                                `https://webapi.tobit.com/TextStringService/v1.0/V2/LangStrings?libName=${TEXTSTRING_CONFIG.libName}`,
                                 {
-                                    method: 'GET',
-                                    cache: 'no-cache',
+                                    method: 'PUT',
                                     headers: new Headers({
-                                        Authorization: `Bearer ${chayns.env.user.tobitAccessToken}`
+                                        Authorization: `Bearer ${chayns.env.user.tobitAccessToken}`,
+                                        'Content-Type': 'application/json'
+                                    }),
+                                    body: JSON.stringify({
+                                        description: '',
+                                        stringName: `${TEXTSTRING_CONFIG.prefix}${stringName}`,
+                                        textEng: '',
+                                        textFra: '',
+                                        textGer: fallback,
+                                        textIt: '',
+                                        textNl: '',
+                                        textES: '',
+                                        textPT: '',
+                                        textTR: '',
+                                        textPL: '',
+                                        toTranslate: ['en', 'nl', 'it', 'fr', 'pt', 'es', 'tr', 'pl']
                                     })
                                 }
                             );
-                            const libContent = await libResponse.json();
-                            if (libResponse.status === 200 && libContent && Array.isArray(libContent)
-                                && !libContent.find(s => s.stringName === `${TEXTSTRING_CONFIG.prefix}${stringName}`)) {
-                                const response = await fetch(
-                                    `https://webapi.tobit.com/TextStringService/v1.0/V2/LangStrings?libName=${TEXTSTRING_CONFIG.libName}`,
-                                    {
-                                        method: 'PUT',
-                                        headers: new Headers({
-                                            Authorization: `Bearer ${chayns.env.user.tobitAccessToken}`,
-                                            'Content-Type': 'application/json'
-                                        }),
-                                        body: JSON.stringify({
-                                            description: '',
-                                            stringName: `${TEXTSTRING_CONFIG.prefix}${stringName}`,
-                                            textEng: '',
-                                            textFra: '',
-                                            textGer: fallback,
-                                            textIt: '',
-                                            textNl: '',
-                                            textES: '',
-                                            textPT: '',
-                                            textTR: '',
-                                            toTranslate: ['en', 'nl', 'it', 'fr', 'pt', 'es', 'tr']
-                                        })
-                                    }
+                            if (response && response.status === 201) {
+                                console.warn(
+                                    `[TextString] Created string '${TEXTSTRING_CONFIG.prefix}${stringName}' as '${fallback}'. Translated to: en, nl, it, fr, pt, es, tr, pl.`
                                 );
-                                if (response && response.status === 201) {
-                                    console.warn(
-                                        `[TextString] Created string '${TEXTSTRING_CONFIG.prefix}${stringName}' as '${fallback}'. Translated to: ${[
-                                            'en',
-                                            'nl',
-                                            'it',
-                                            'fr',
-                                            'pt',
-                                            'es',
-                                            'tr'
-                                        ].join(', ')}.`
-                                    );
-                                }
                             }
-                        });
-                    }
+                        }
+                    });
                 }
             } catch (e) {
                 // ignored
@@ -187,7 +165,7 @@ const TextStringReplacer: FunctionComponent<TextStringReplacerConfig> = ({
         : calculatedString;
 
     // generate a guid used for react keys
-    const guid = useMemo(() => generateUUID(), []);
+    const guid = useMemo(() => generateGuid(), []);
 
     // calculate the actual content with replacements. To display a mix of strings and react elements this function
     // creates an array of strings and react elements that is split further and further the more jsx replacements occur